Update deployment.yaml
brute force the storage path to the 100 user (apache). Seems like 1000 is an error, and this will also take care of the log warning. Will test this, and then likely need to create a root variable that handles the path, and then let the sessions path just branch straight off of that.